Guest page splits across Onboarding, Users and Gdpr (nobodies-collective/Humans#1091)
Shell's GuestController splits by ownership, routes byte-identical:
- Humans.Onboarding: /Guest dashboard frame + GuestDashboardViewModel + Guest_* resx keys
- Humans.Users: GuestAccountController (comms prefs incl. utoken mode, deletion request/cancel) + CommunicationPreferencesPanel component
- Humans.Gdpr: GuestDataController (/Guest/DownloadData) calling its own IGdprExportService
