+++ src/Humans.Base/Resources/SharedResource.de.resx
@@ -460,75 +460,6 @@
   <data name="Dashboard_DeclareNotAttending" xml:space="preserve">
     <value>Komme dieses Jahr nicht</value>
   </data>
-  <data name="Guest_Greeting" xml:space="preserve">
-    <value>Hey, {0}.</value>
-  </data>
-  <data name="Guest_Subtitle" xml:space="preserve">
-    <value>Willkommen bei Humans — wo Nobodies Collective alles für Elsewhere verwaltet.</value>
-  </data>
-  <data name="Guest_GetInvolvedTitle" xml:space="preserve">
-    <value>Bereit, kein Geist mehr zu sein?</value>
-  </data>
-  <data name="Guest_GetInvolvedDescription" xml:space="preserve">
-    <value>Du existierst gerade… aber niemand kann dich sehen. Erstelle dein Profil und plötzlich:</value>
-  </data>
-  <data name="Guest_GetInvolvedBullet1" xml:space="preserve">
-    <value>Du kannst Teams beitreten</value>
-  </data>
-  <data name="Guest_GetInvolvedBullet2" xml:space="preserve">
-    <value>Du kannst Schichten übernehmen</value>
-  </data>
-  <data name="Guest_GetInvolvedBullet3" xml:space="preserve">
-    <value>Leute wissen, wer du bist</value>
-  </data>
-  <data name="Guest_GetInvolvedHint" xml:space="preserve">
-    <value>Dauert ~2 Minuten.</value>
-  </data>
-  <data name="Guest_CreateProfile" xml:space="preserve">
-    <value>Mein Profil erstellen</value>
-  </data>
-  <data name="Guest_Explore" xml:space="preserve">
-    <value>Schau dich um</value>
-  </data>
-  <data name="Guest_ExploreSubtitle" xml:space="preserve">
-    <value>Kein Profil nötig. Nur Neugier.</value>
-  </data>
-  <data name="Guest_BarriosDescription" xml:space="preserve">
-    <value>Die Nachbarschaften, die Leute während des Events bauen und bewohnen</value>
-  </data>
-  <data name="Guest_TeamsDescription" xml:space="preserve">
-    <value>Die Crews, die alles am Laufen halten — Sound, Sicherheit, Aufbau, alles</value>
-  </data>
-  <data name="Guest_Legal" xml:space="preserve">
-    <value>Rechtliches</value>
-  </data>
-  <data name="Guest_LegalDescription" xml:space="preserve">
-    <value>Datenschutz, Bedingungen, Einwilligungen — langweilig aber wichtig</value>
-  </data>
-  <data name="Guest_AccountTitle" xml:space="preserve">
-    <value>Dein Kram</value>
-  </data>
-  <data name="Guest_AccountSubtitle" xml:space="preserve">
-    <value>Auch ohne Profil kannst du dein Konto und deine Daten verwalten.</value>
-  </data>
-  <data name="Guest_CommsTitle" xml:space="preserve">
-    <value>Kommunikation</value>
-  </data>
-  <data name="Guest_CommsDescription" xml:space="preserve">
-    <value>Wähle, welche E-Mails und Benachrichtigungen du von uns bekommst</value>
-  </data>
-  <data name="Guest_DataTitle" xml:space="preserve">
-    <value>Deine Daten</value>
-  </data>
-  <data name="Guest_DataDescription" xml:space="preserve">
-    <value>Lade alles herunter, was wir über dich haben — es gehört dir</value>
-  </data>
-  <data name="Guest_DeleteTitle" xml:space="preserve">
-    <value>Verschwinden</value>
-  </data>
-  <data name="Guest_DeleteDescription" xml:space="preserve">
-    <value>Kontolöschung beantragen — 30 Tage, um es dir anders zu überlegen</value>
-  </data>
   <data name="Guest_TicketsTitle" xml:space="preserve">
     <value>Deine Tickets</value>
   </data>
@@ -546,21 +477,6 @@
   </data>
   <data name="Guest_TicketsConfirmed" xml:space="preserve">
     <value>Du hast ein Ticket für das Event.</value>
-  </data>
-  <data name="Guest_DeletionPendingTitle" xml:space="preserve">
-    <value>Du bist auf dem Weg raus</value>
-  </data>
-  <data name="Guest_DeletionRequestedOn" xml:space="preserve">
-    <value>Angefragt am {0}.</value>
-  </data>
-  <data name="Guest_DeletionHeldUntil" xml:space="preserve">
-    <value>Wir löschen alles nach dem {0} — warten bis nach dem Event.</value>
-  </data>
-  <data name="Guest_DeletionScheduledFor" xml:space="preserve">
-    <value>Alles weg am {0}. Wirklich.</value>
-  </data>
-  <data name="Guest_DeletionCancelButton" xml:space="preserve">
-    <value>Hab's mir anders überlegt</value>
   </data>
   <data name="Privacy_Title" xml:space="preserve">
     <value>Datenschutzrichtlinie</value>
