+++ src/Humans.Base/Resources/SharedResource.it.resx
@@ -460,75 +460,6 @@
   <data name="Dashboard_DeclareNotAttending" xml:space="preserve">
     <value>Non vengo quest'anno</value>
   </data>
-  <data name="Guest_Greeting" xml:space="preserve">
-    <value>Ciao, {0}.</value>
-  </data>
-  <data name="Guest_Subtitle" xml:space="preserve">
-    <value>Benvenuto/a su Humans — dove Nobodies Collective gestisce tutto per Elsewhere.</value>
-  </data>
-  <data name="Guest_GetInvolvedTitle" xml:space="preserve">
-    <value>Pronto/a a smettere di essere un fantasma?</value>
-  </data>
-  <data name="Guest_GetInvolvedDescription" xml:space="preserve">
-    <value>Ora esisti… ma nessuno ti vede. Crea il tuo profilo e improvvisamente:</value>
-  </data>
-  <data name="Guest_GetInvolvedBullet1" xml:space="preserve">
-    <value>Puoi unirti ai team</value>
-  </data>
-  <data name="Guest_GetInvolvedBullet2" xml:space="preserve">
-    <value>Puoi prendere turni</value>
-  </data>
-  <data name="Guest_GetInvolvedBullet3" xml:space="preserve">
-    <value>La gente sa chi sei</value>
-  </data>
-  <data name="Guest_GetInvolvedHint" xml:space="preserve">
-    <value>~2 minuti.</value>
-  </data>
-  <data name="Guest_CreateProfile" xml:space="preserve">
-    <value>Crea il mio profilo</value>
-  </data>
-  <data name="Guest_Explore" xml:space="preserve">
-    <value>Dai un'occhiata</value>
-  </data>
-  <data name="Guest_ExploreSubtitle" xml:space="preserve">
-    <value>Non serve un profilo. Solo curiosità.</value>
-  </data>
-  <data name="Guest_BarriosDescription" xml:space="preserve">
-    <value>I quartieri che la gente costruisce e abita durante l'evento</value>
-  </data>
-  <data name="Guest_TeamsDescription" xml:space="preserve">
-    <value>I team che fanno funzionare tutto — audio, sicurezza, costruzioni, tutto</value>
-  </data>
-  <data name="Guest_Legal" xml:space="preserve">
-    <value>Legale</value>
-  </data>
-  <data name="Guest_LegalDescription" xml:space="preserve">
-    <value>Privacy, termini, consensi — noioso ma importante</value>
-  </data>
-  <data name="Guest_AccountTitle" xml:space="preserve">
-    <value>Le tue cose</value>
-  </data>
-  <data name="Guest_AccountSubtitle" xml:space="preserve">
-    <value>Anche senza profilo, puoi gestire il tuo account e i tuoi dati.</value>
-  </data>
-  <data name="Guest_CommsTitle" xml:space="preserve">
-    <value>Comunicazioni</value>
-  </data>
-  <data name="Guest_CommsDescription" xml:space="preserve">
-    <value>Scegli quali email e notifiche ricevere da noi</value>
-  </data>
-  <data name="Guest_DataTitle" xml:space="preserve">
-    <value>I tuoi dati</value>
-  </data>
-  <data name="Guest_DataDescription" xml:space="preserve">
-    <value>Scarica tutto quello che abbiamo su di te — è tuo</value>
-  </data>
-  <data name="Guest_DeleteTitle" xml:space="preserve">
-    <value>Sparire</value>
-  </data>
-  <data name="Guest_DeleteDescription" xml:space="preserve">
-    <value>Richiedi la cancellazione dell'account — 30 giorni per cambiare idea</value>
-  </data>
   <data name="Guest_TicketsTitle" xml:space="preserve">
     <value>I tuoi biglietti</value>
   </data>
@@ -546,21 +477,6 @@
   </data>
   <data name="Guest_TicketsConfirmed" xml:space="preserve">
     <value>Hai un biglietto per l'evento.</value>
-  </data>
-  <data name="Guest_DeletionPendingTitle" xml:space="preserve">
-    <value>Stai uscendo</value>
-  </data>
-  <data name="Guest_DeletionRequestedOn" xml:space="preserve">
-    <value>Richiesto il {0}.</value>
-  </data>
-  <data name="Guest_DeletionHeldUntil" xml:space="preserve">
-    <value>Cancelleremo tutto dopo il {0} — aspettiamo la fine dell'evento.</value>
-  </data>
-  <data name="Guest_DeletionScheduledFor" xml:space="preserve">
-    <value>Tutto sparisce il {0}. Sul serio.</value>
-  </data>
-  <data name="Guest_DeletionCancelButton" xml:space="preserve">
-    <value>Ho cambiato idea</value>
   </data>
   <data name="Privacy_Title" xml:space="preserve">
     <value>Informativa sulla Privacy</value>
