+++ src/Sections/Humans.Debug/Views/WidgetGallery/Index.cshtml
@@ -329,7 +329,7 @@
                             <span class="wg-card-name">&lt;vc:communication-preferences-panel&gt;</span>
                             <span class="wg-card-kind kind-vc">ViewComponent</span>
                         </div>
-                        <span class="wg-card-path">src/Humans.Web/ViewComponents/CommunicationPreferencesPanelViewComponent.cs</span>
+                        <span class="wg-card-path">src/Sections/Humans.Users/ViewComponents/CommunicationPreferencesPanelViewComponent.cs</span>
                     </div>
                     <div class="wg-card-note">
                         Per-user communication-preferences matrix. Write mode (default) renders toggle checkboxes that POST to <code>/Profile/Me/CommunicationPreferences/Update</code>. Read-only mode hides the form and surfaces <code>UpdateSource</code> + <code>UpdatedAt</code> per category — used on <code>/Users/Admin/{id}</code>.
