+++ src/Sections/Humans.Onboarding.Contracts/IOnboardingWidgetState.cs
@@ -7,9 +7,9 @@ namespace Humans.Onboarding.Contracts;
 /// No new tables; no new claims.
 /// </summary>
 /// <remarks>
-/// On the leaf because Shell's <c>GuestController</c> bounces a mid-widget user back
-/// into the flow with it; the section's own dispatcher and progress banner are the
-/// other two consumers.
+/// On the leaf because this section's own <c>GuestController</c> bounces a mid-widget
+/// user back into the flow with it; the section's own dispatcher and progress banner
+/// are the other two consumers.
 /// </remarks>
 public interface IOnboardingWidgetState
 {
