+++ src/Sections/Humans.Onboarding/OnboardingResource.resx
@@ -111,4 +111,32 @@
   <data name="Welcome_CTA" xml:space="preserve"><value>Sign in or create account</value></data>
   <data name="Welcome_Footer" xml:space="preserve"><value>Already have an account? Same button — sign in and you'll land straight on the shifts page.</value></data>
   <data name="Welcome_Contact" xml:space="preserve"><value>Questions? Reach us at</value></data>
+  <data name="Guest_AccountSubtitle" xml:space="preserve"><value>Even without a profile, you can manage your account and data.</value></data>
+  <data name="Guest_AccountTitle" xml:space="preserve"><value>Your stuff</value></data>
+  <data name="Guest_BarriosDescription" xml:space="preserve"><value>The neighbourhoods people build and live in during the event</value></data>
+  <data name="Guest_CommsDescription" xml:space="preserve"><value>Choose which emails and notifications you get from us</value></data>
+  <data name="Guest_CommsTitle" xml:space="preserve"><value>Comms</value></data>
+  <data name="Guest_CreateProfile" xml:space="preserve"><value>Create my profile</value></data>
+  <data name="Guest_DataDescription" xml:space="preserve"><value>Download everything we have on you — it's yours</value></data>
+  <data name="Guest_DataTitle" xml:space="preserve"><value>Your data</value></data>
+  <data name="Guest_DeleteDescription" xml:space="preserve"><value>Request full account deletion — 30 days to change your mind</value></data>
+  <data name="Guest_DeleteTitle" xml:space="preserve"><value>Disappear</value></data>
+  <data name="Guest_DeletionCancelButton" xml:space="preserve"><value>Changed my mind</value></data>
+  <data name="Guest_DeletionHeldUntil" xml:space="preserve"><value>We'll delete everything after {0} — waiting until after the event.</value></data>
+  <data name="Guest_DeletionPendingTitle" xml:space="preserve"><value>You're on your way out</value></data>
+  <data name="Guest_DeletionRequestedOn" xml:space="preserve"><value>Requested on {0}.</value></data>
+  <data name="Guest_DeletionScheduledFor" xml:space="preserve"><value>Everything goes on {0}. For real.</value></data>
+  <data name="Guest_Explore" xml:space="preserve"><value>Look around</value></data>
+  <data name="Guest_ExploreSubtitle" xml:space="preserve"><value>No profile needed. Just curiosity.</value></data>
+  <data name="Guest_GetInvolvedBullet1" xml:space="preserve"><value>You can join teams</value></data>
+  <data name="Guest_GetInvolvedBullet2" xml:space="preserve"><value>You can grab shifts</value></data>
+  <data name="Guest_GetInvolvedBullet3" xml:space="preserve"><value>People actually know who you are</value></data>
+  <data name="Guest_GetInvolvedDescription" xml:space="preserve"><value>Right now you exist… but no one can see you. Create your profile and suddenly:</value></data>
+  <data name="Guest_GetInvolvedHint" xml:space="preserve"><value>Takes ~2 minutes.</value></data>
+  <data name="Guest_GetInvolvedTitle" xml:space="preserve"><value>Ready to stop being a ghost?</value></data>
+  <data name="Guest_Greeting" xml:space="preserve"><value>Hey, {0}.</value></data>
+  <data name="Guest_Legal" xml:space="preserve"><value>Legal</value></data>
+  <data name="Guest_LegalDescription" xml:space="preserve"><value>Privacy policy, terms, consent docs — the boring but important stuff</value></data>
+  <data name="Guest_Subtitle" xml:space="preserve"><value>Welcome to Humans — where Nobodies Collective manages everything for Elsewhere.</value></data>
+  <data name="Guest_TeamsDescription" xml:space="preserve"><value>The crews that keep things running — sound, safety, builds, all of it</value></data>
 </root>
