+++ src/Sections/Humans.Users/Views/GuestAccount/_ViewImports.cshtml
@@ -1,0 +1,31 @@
+@* A section RCL does NOT inherit the host's Views/_ViewImports.cshtml. Omitting a line
+   here ships broken markup with a green build and no runtime error — see design §1/§2.
+   Folder-level rather than one Views/_ViewImports.cshtml because Humans.Users still grants
+   InternalsVisibleTo to Humans.Web: a root _ViewImports in both assemblies generates the
+   same internal Views__ViewImports type and Shell fails with CS0436. Collapse these into
+   one root file in the commit that drops the grant. *@
+@using Humans.Base.Extensions
+@using Humans.Base.Enums
+@using Humans.Users
+@using Humans.Users.Contracts
+@using Humans.Users.Controllers
+@using Humans.Users.Models
+@using Humans.Users.Services
+@using Humans.Base
+@using Humans.Base.Authorization
+@using Humans.Base.Helpers
+@using Humans.Base.Models
+@using Humans.Base.Models.Tables
+@using Humans.Base.ViewComponents
+@using Microsoft.AspNetCore.Authorization
+@using Microsoft.AspNetCore.Html
+@using Microsoft.Extensions.Localization
+@using NodaTime
+@* Users' ~250 keys are still in Humans.UI's SharedResource — the resx carve into
+   UsersResource has not happened yet, so this binds to the shared set exactly as the
+   Shell _ViewImports these views came from did. Rebind when UsersResource lands. *@
+@inject IStringLocalizer<SharedResource> Localizer
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
+@* <vc:human> and the other shared components are Humans.UI's. *@
+@addTagHelper *, Humans.Base
+@addTagHelper *, Humans.Users
