+++ src/Sections/Humans.Users/Views/Shared/Components/CommunicationPreferencesPanel/Default.cshtml
@@ -1,5 +1,5 @@
 @using Humans.Base.Extensions
-@model Humans.Web.Models.CommunicationPreferencesViewModel
+@model Humans.Users.Models.CommunicationPreferencesViewModel
 @{
     var isReadOnly = Model.ReadOnly;
 }
